requirements.txt
Django==2.0.13
django-crispy-forms==1.8.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.0.0
psycopg2==2.7.7
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
